backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
anthropic==0.64.0
openai==1.99.9
pillow==11.3.0
pillow-heif==0.16.0